Turn on W4

+++ src/doom/info.cpp
@@ -1259,7 +1259,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                           // mass
         0,                                             // damage
         sfx_None,                                      // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                         // raisestate
     },
 
@@ -1340,7 +1340,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                           // mass
         0,                                             // damage
         sfx_None,                                      // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                         // raisestate
     },
 
@@ -1394,7 +1394,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         8,                                                                       // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -1583,7 +1583,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         8,                                                                       // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -1934,7 +1934,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                           // mass
         0,                                             // damage
         sfx_None,                                      // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                         // raisestate
     },
 
@@ -1988,7 +1988,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         3,                                                                       // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -2015,7 +2015,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         5,                                                                       // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -2069,7 +2069,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         5,                                                                       // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -2096,7 +2096,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         100,                                                                     // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -2123,7 +2123,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                                     // mass
         5,                                                                       // damage
         sfx_None,                                                                // activesound
-        MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_MISSILE | MF_DROPOFF | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                                                   // raisestate
     },
 
@@ -2150,7 +2150,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                                          // mass
         0,                                                            // damage
         sfx_None,                                                     // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_FLIPPABLE | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_FLIPPABLE | MF_TRANSLUCENT), // flags
         S_NULL                                                        // raisestate
     },
 
@@ -2204,7 +2204,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                           // mass
         0,                                             // damage
         sfx_None,                                      // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                         // raisestate
     },
 
@@ -2231,7 +2231,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                           // mass
         0,                                             // damage
         sfx_None,                                      // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                         // raisestate
     },
 
@@ -2285,7 +2285,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                           // mass
         0,                                             // damage
         sfx_None,                                      // activesound
-        MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP | MF_NOGRAVITY | MF_TRANSLUCENT), // flags
         S_NULL                                         // raisestate
     },
 
@@ -2636,7 +2636,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                        // mass
         0,                                          // damage
         sfx_None,                                   // activesound
-        MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT), // flags
         S_NULL                                      // raisestate
     },
 
@@ -2663,7 +2663,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                        // mass
         0,                                          // damage
         sfx_None,                                   // activesound
-        MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT), // flags
         S_NULL                                      // raisestate
     },
 
@@ -2717,7 +2717,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                        // mass
         0,                                          // damage
         sfx_None,                                   // activesound
-        MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT), // flags
         S_NULL                                      // raisestate
     },
 
@@ -2825,7 +2825,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] = {
         100,                                        // mass
         0,                                          // damage
         sfx_None,                                   // activesound
-        MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT, // flags
+        static_cast<unsigned int>(MF_SPECIAL | MF_COUNTITEM | MF_TRANSLUCENT), // flags
         S_NULL                                      // raisestate
     },
