+++ src/strife/am_map.cpp
@@ -1044,7 +1044,7 @@ AM_drawFline
 	d = ay - ax/2;
 	while (1)
 	{
-	    PUTDOT(x,y,color);
+	    PUTDOT(x,y,static_cast<uint8_t>(color));
 	    if (x == fl->b.x) return;
 	    if (d>=0)
 	    {
@@ -1060,7 +1060,7 @@ AM_drawFline
 	d = ax - ay/2;
 	while (1)
 	{
-	    PUTDOT(x, y, color);
+	    PUTDOT(x, y, static_cast<uint8_t>(color));
 	    if (y == fl->b.y) return;
 	    if (d >= 0)
 	    {
